Revisão do codigo, finalização da estilização e criação da tela de login

+++ src/pages/home/home.js
@@ -5,7 +5,6 @@ import RemoveButton from '../../componentes/removeButton/removeButton';
 import Header from '../../componentes/header/header';
 import Footer from '../../componentes/footer/footer';
 import CategoriaFiltro from '../../componentes/filtro/filtro';
-import DownloadJsonButton from '../../componentes/downloadJson/downloadJson';
 
 function Home() {
   const [produtos, setProdutos] = useState(() => {
@@ -20,12 +19,30 @@ function Home() {
     navigate('/cadastro');
   };
 
-  const handleRemove = (id) => {
-    const produtosAtualizados = produtos.filter((_, index) => index !== id);
-    setProdutos(produtosAtualizados);
-    localStorage.setItem('produtos', JSON.stringify(produtosAtualizados));  // Atualiza o localStorage após remover um item
+  const formatarDataParaExibicao = (dataISO) => {
+    const [ano, mes, dia] = dataISO.split('-');
+    return `${dia}/${mes}/${ano}`;
   };
-  
+
+  const converterDataParaISO = (dataBrasileira) => {
+    const [dia, mes, ano] = dataBrasileira.split('/');
+    return `${ano}-${mes}-${dia}`;
+  };
+
+  const ordenarPorDataDeValidade = (produtos) => {
+    return produtos.sort((a, b) => new Date(converterDataParaISO(a.validade)) - new Date(converterDataParaISO(b.validade)));
+  };
+
+  const handleRemove = (id) => {
+    const removee = window.confirm('Desseja remover esse item?')
+
+    if (removee) {
+      const produtosAtualizados = produtos.filter((_, index) => index !== id);
+      setProdutos(produtosAtualizados);
+      localStorage.setItem('produtos', JSON.stringify(produtosAtualizados));
+    }
+     
+  };
 
   useEffect(() => {
     const onStorageUpdate = () => {
@@ -39,46 +56,40 @@ function Home() {
     return () => window.removeEventListener('storage', onStorageUpdate);
   }, []);
   
+  const produtoContainerClass = filtroCategoria === '' ? "flex-row" : "flex-col";
 
   return (
     <div className="relative mb-[200px]">
       <Header />
       <div className='flex flex-col justify-center items-center gap-[20px] mt-[20px] mb-[20px]'>
         <AdicionarBotao onClick={handleAdicionarClick} />
-
         <CategoriaFiltro filtroCategoria={filtroCategoria} setFiltroCategoria={setFiltroCategoria} />
       </div>
     
       <hr />
 
-      <div className="flex flex-row justify-center items-center gap-8 flex-wrap mt-5 z-0 text-slate-200 font-semibold">
-
-        {produtos.filter(produto => filtroCategoria === '' || produto.categoria === filtroCategoria).map((produto, index) => (
-
+      <div className={`flex justify-center items-center gap-8 flex-wrap mt-5 z-0 text-slate-200 font-semibold ${produtoContainerClass}`}>
+        {ordenarPorDataDeValidade(produtos.filter(produto => filtroCategoria === '' || produto.categoria === filtroCategoria)).map((produto, index) => (
             <div key={index} className="relative flex items-start bg-[rgb(107,109,109)] w-[600px] h-auto rounded-[10px]">
-
               <div className="flex items-center justify-center bg-white rounded-md w-[120px] h-[120px] min-w-[120px] mx-3 my-3">
                 {produto.imagemBase64 && <img className="w-auto h-[100px]" src={produto.imagemBase64} alt={`Foto do ${produto.nomeProduto}`} />}
               </div>
 
               <div className='mt-[10px]'>
                 <h3 className='font-bold text-[18px] text-white'>{produto.nomeProduto}</h3>
-                <p>Data de entrada: {produto.dataEntrada}</p>
-                <p>Validade: {produto.validade}</p>
+                <p>Data de entrada: {formatarDataParaExibicao(produto.dataEntrada)}</p>
+                <p>Validade: {formatarDataParaExibicao(produto.validade)}</p>
                 <p>Categoria: {produto.categoria}</p>
               </div>
 
               <div className="absolute right-[5px] top-[5px]">
                 <RemoveButton onRemove={() => handleRemove(index)} id={index} />
               </div>
-
             </div>
           ))}
       </div>
-
       <div className="fixed left-0 bottom-0 w-screen h-[70px] z-10">
-        <DownloadJsonButton></DownloadJsonButton>
-
+        
         <Footer />
       </div>
     </div>
